requirements.txt
google-cloud-storage==2.1.0
google-cloud-bigquery==3.1.0
pandas==1.4.2